Add test for edid parsing and fix reading gamma
Summary:
The bug in reading the gamma value was there from the start (2012),
dividing the number by 100 will result in a truncated value.
Add test so that we can verify that the parsing of the EDIDs work,
it will be easy to add other problematic EDIDs later on.

Let edidDecodeFraction return float

All uses of the function are QQuaternion setters which take floats.
This saves a bunch of conversions and should thus be faster. The edid
test shows that the result is unchanged in all digits.

Reviewers: #plasma, romangg

Reviewed By: #plasma, romangg

Subscribers: romangg, plasma-devel

Tags: #plasma

Differential Revision: https://phabricator.kde.org/D14418

+++ src/edid.cpp
@@ -308,7 +308,7 @@ bool Edid::Private::parse(const QByteArray &rawData)
     if (data[GCM_EDID_OFFSET_GAMMA] == 0xff) {
         gamma = 1.0;
     } else {
-        gamma = (static_cast<float>(data[GCM_EDID_OFFSET_GAMMA] / 100) + 1);
+        gamma = data[GCM_EDID_OFFSET_GAMMA] / 100.0 + 1.0;
     }
 
     /* get color red */
